Added *.bak file check to wmp_local_server_config.sh and temporarily change EC2 server information
WARNING: local_server_config.sh and wmp_server_config.sh were renamed to wmp_local_server_config.sh and wmp_ec2_server_config.sh because both were made accessable from everywhere in the filesystem which raised the need for a more specific naming convention.

Previously wmp_local_server_config.sh removed the XML setting files also if there were no bak files to replace them. We now check for the existence of the bak files before deleting the XMLs.

We also temporarily changed the XMLs for the EC2 server so that we can run the Mobilis server locally and it'll connect to the EC2 XMPP server. Note, that you may also run this configuration on the EC2 server, but it will be slower than the localhost configuration as there has to be made a DNS lookup first before being able to communicate with the XMPP server.

+++ MobilisServer/wmp_server_config/wmp_local_server_config.sh
@@ -1,7 +1,15 @@
-#!/bin/bash
+#usr/bin/sh
 # This scripts overrides the current Mobilis Server configuration as defined in /src/META-INF/MobilisSettings.xml and hibernate.xml by the default local configuration saved in the appropriate bak files by the a accompanied wmp_server_config.sh script. This script is normally run *after* the wmp_server_config.sh script and after a runnable jar file has been created for deployment on the Amazon EC2 server. It returns the Server project back to normal development state.
 # @sven 06.2011
 
+if [ ! -f ../src/META-INF/hibernate.xml.bak ]; then 
+	echo "Couldn't find hibernate.xml.bak (normally this means that local config is already in place)- aborting...";
+	exit 1;
+fi
+if [ ! -f ../src/META-INF/MobilisSettings.xml.bak ]; then 
+	echo "Couldn't find MobilisSettings.xml.bak (normally this means that local config is already in place) - aborting..."; 
+	exit 1; 
+fi
 rm ../src/META-INF/hibernate.xml
 rm ../src/META-INF/MobilisSettings.xml
 mv ../src/META-INF/hibernate.xml.bak ../src/META-INF/hibernate.xml
